Remove pkgversions from ppa script.
This version of the ppa script is not yet in production.

The new tarball job is directly triggered by gerrit.  It should
trigger a new version of the PPA job, and this should only ever
happen once per change, so we don't need to keep track of build
numbers in the versions bzr repository.

We do lose the ability to re-run a PPA job when the packaging
repo changes and get a new package version, but since this is
just used for per-commit builds, we can probably wait until the
next auto-build from the next commit.

Change-Id: Idb30d8a73e7a189d00a972c2b9b20417cb090240