Add run-tox.sh
If a bundle file is present, call tox with the jenkins version of
the test environment so it is used.  Otherwise, use the normal
(non-bundle) test environment.  Also, run pip freeze on the
resulting environment at the end so that we have a record of exactly
what packages we ended up testing.

Change-Id: I58af68fece4b3da5f2f2db1dc2fbca60939ce2aa